test-driver/pyproject.toml
[project]
name = "clawbench-test-driver"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["fpdf2>=2.8", "huggingface_hub>=0.27", "pyyaml>=6.0"]
